requirements-docker.txt
alembic==1.16.5
annotated-types==0.7.0
anyio==4.10.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
cryptography==46.0.1
distro==1.9.0
fastapi==0.116.2
greenlet==3.2.4
grpcio==1.75.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
jiter==0.11.0
Mako==1.3.10
MarkupSafe==3.0.2
numpy==2.3.3
google-generativeai==0.8.3
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.3.0
portalocker==3.2.0
protobuf>=3.20.2,<6.0.0
psycopg2-binary==2.9.10
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
pypdfium2==4.30.0
python-dotenv==1.1.1
python-multipart==0.0.20
qdrant-client==1.15.1
requests==2.32.5
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.48.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.35.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
email-validator==2.2.0
PyMuPDF==1.24.14
